POSH, Riker, DPT, LLM benchmarks, automation and containment scripts (#59)
* minor fixes

* added #! and minor fixes

* merged vps

* removed unused scripts from script-globs and .github

* add benchmark_shell on riker run

* merging dynamic analysis

* more missing #!

* fixes for dynamic analysis

* added time on dockerimage + rename to koala

* changed naming to koala, added txt output to viz/dynamic.py

* added koala on readme

* removed ascii koala

* changed permissions in infra scripts

* changed run_dynamic so it can create  .out and .err. added dynamic analysis to main.sh through --resources flag (only for docker) and added text output to viz/dynamic.py

* made bio cleanup better

* run all script

* changed infra scripts, csvs and jsons for vps-audit merge

* dynamic analysis mostly working

* automation scripts

* Dockerfiles for all scripts

* formatted json

* made scripts in commit executable

* fix run vpsaudit

* added gitignores

* added backup process logs to gitignore

* added to dockerfile

* added hashes and min inputs in aurpkg

* create_docker_image.sh

* added script to create docker images

* added script to create docker images

* opt-parallel

* prepare bio for min inps

* prep bio small inputs

* m

* script to create docker images for each script

* check bio samtools version

* minor safety fixes in aurpkg - standardizing format

* bio - standardizing format

* standardizing format

* standardizing format, filenames etc.

* standardizing format, filenames etc.

* fix makeself

* standardizing format

* fixes

* added todos for min versions

* fixes

* removed some files

* fixes

* fixes

* fixed file enc deps

* git workflow benchmark

* git-workflwo bench

* ray-tracing posh benchmark

* minor fix

* poch mogrify bench

* riker-lsof

* added 2 riker benchmarks minor changes to posh

* changes to posh

* fix deps

* hashes

* migrating added

* fix

* verification changed

* verification

* completed riker benchmarks

* added llm image annotation benchmark

* fixes - image annotation benchmark

* removed tmp

* fix for small and min

* main fix

* llm image annotation benchmark

* fix

* riker fix

* riker fix

* fix image annotation deps

* fix

* fix

* fix img annot

* posh

* added min inputs to aurpkg bio covid

* added min inputs file enc

* added min inputs

* standardize locale

* added min inputs for media-conv max-temp image-annotation

* posh git workflow

* min inputs log analysis

* min inputs max-temp

* min inputs

* fix min/small inputs

* ray tracing inputs

* completed posh

* fix

* media-conv: build imagemagick

* fix media conv

* changed ci

* fix

* fix ci

* added new benchmarks in ci

* fixes

* more fixes + changing ci

* fix

* fix

* fix

* fix

* fixed verification port scan

* fixes and adding opt parallel to ci

* last fix

* minor changes

* changes in verification + renaming

* tuft-weather

* removed tuft weather

* fix

* fix input.sh dpt

* dpt verification + fixes

* minor fixes

* typo fix riker

* try alliases in git workflow

* fixes dpt verification

* added locale override for ls to work in a standardized way

* fix web-index

* removed dockerimage

* renamed image annotation

* minor change

* ci test dpt

* ci test dpt

* ci test dpt

* ci test dpt

* removed changes

* fixed verification

* fix

* initial commit for playlist-creation

* added input checks

* fix

* added autoconf

* added multiple run option, result aggregation and tutorial in readme

* merged  tutorial and dyn analysis

* fixes

* fix

* fix

* readme change

* unquoted  to parse shell arguments

* fix run_all benchmark shell passing

* changed time-run file

+++ media-conv/input.sh
@@ -2,38 +2,77 @@
 
 REPO_TOP=$(git rev-parse --show-toplevel)
 eval_dir="${REPO_TOP}/media-conv"
-input_dir="${eval_dir}/input"
+input_dir="${eval_dir}/inputs"
 
-mkdir -p $input_dir
+mkdir -p "$input_dir"
 
 data_url=https://atlas-group.cs.brown.edu/data/wav.zip
 zip_dst=$input_dir/wav.zip
 full_dir="$input_dir/wav_full"
 small_dir="$input_dir/wav_small"
-wget --no-check-certificate "$data_url" -O "$zip_dst"
-unzip "$zip_dst" -d "$input_dir"
-mkdir -p "$full_dir" "$small_dir"
-# copy `.wav`s to their final destinations. 
-# Make sure we have the correct number of inputs
-# with numbered backups (do not overwrite inputs).
-for i in {1..120}; do
-    cp --backup=numbered $input_dir/wav/* "--target-directory=$full_dir"
-done
-for i in {1..10}; do
-    cp --backup=numbered $input_dir/wav/* "--target-directory=$small_dir"
-done
-rm -r "$zip_dst" "$input_dir/wav"
+min_dir="$input_dir/wav_min"
 
+if [[ ! -d "$full_dir" ]]; then
+    wget --no-check-certificate "$data_url" -O "$zip_dst"
+    unzip "$zip_dst" -d "$input_dir"
+    rm -rf "$full_dir" "$small_dir" "$min_dir"
+    mkdir -p "$full_dir" "$small_dir" "$min_dir"
+
+    # copy `.wav`s to their final destinations.
+    # Make sure we have the correct number of inputs
+    # with numbered backups (do not overwrite inputs).
+    for i in {1..120}; do
+        cp --backup=numbered $input_dir/wav/* "--target-directory=$full_dir"
+    done
+    for i in {1..10}; do
+        cp --backup=numbered $input_dir/wav/* "--target-directory=$small_dir"
+    done
+    for i in {1..2}; do
+        cp --backup=numbered $input_dir/wav/* "--target-directory=$min_dir"
+    done
+    rm -r "$zip_dst" "$input_dir/wav"
+fi
+
+# if small flag
+for arg in "$@"; do
+    if [[ "$arg" == "--small" ]]; then
+        if [[ -d "$input_dir/jpg_small" ]]; then
+            echo "Data already downloaded and extracted."
+            exit 0
+        fi
+        data_url=https://atlas-group.cs.brown.edu/data/small/jpg.zip
+        zip_dst=$input_dir/jpg_small.zip
+        out_dir=$input_dir/jpg_small
+        wget --no-check-certificate $data_url -O $zip_dst || {
+            echo "Failed to download $data_url"
+            exit 1
+        }
+        unzip $zip_dst -d $out_dir || {
+            echo "Failed to unzip $zip_dst"
+            exit 1
+        }
+        rm "$zip_dst"
+        exit 0
+    elif [[ "$arg" == "--min" ]]; then
+        #TODO min versions
+        if [[ -d "$input_dir/jpg_min" ]]; then
+            echo "Data already downloaded and extracted."
+            exit 0
+        fi
+        min_inputs="$eval_dir/min_inputs/"
+        mkdir -p "$input_dir"
+        cp -r "$min_inputs"/* "$input_dir/"
+        exit 0
+    fi
+done
+
+if [[ -d "$input_dir/jpg" ]]; then
+    echo "Data already downloaded and extracted."
+    exit 0
+fi
 data_url=https://atlas-group.cs.brown.edu/data/full/jpg.zip
 zip_dst="$input_dir/jpg_full.zip"
 out_dir="$input_dir/jpg_full"
-wget --no-check-certificate $data_url -O $zip_dst
-unzip $zip_dst -d $out_dir
-rm "$zip_dst"
-
-data_url=https://atlas-group.cs.brown.edu/data/small/jpg.zip
-zip_dst=$input_dir/jpg_small.zip
-out_dir=$input_dir/jpg_small
 wget --no-check-certificate $data_url -O $zip_dst
 unzip $zip_dst -d $out_dir
 rm "$zip_dst"
